Move matrix based Stokes solver to own class
