+++ source/simulator/core.cc
@@ -27,6 +27,7 @@
 #include <aspect/advection_field.h>
 #include <aspect/volume_of_fluid/handler.h>
 #include <aspect/newton.h>
+#include <aspect/simulator/solver/stokes_matrix_based.h>
 #include <aspect/simulator/solver/stokes_matrix_free.h>
 #include <aspect/simulator/solver/stokes_direct.h>
 #include <aspect/mesh_deformation/interface.h>
@@ -462,18 +463,23 @@ namespace aspect
     if (parameters.stokes_solver_type == Parameters<dim>::StokesSolverType::block_gmg)
       {
         stokes_matrix_free = create_matrix_free_solver<dim>(*this, parameters);
-
         stokes_matrix_free->initialize_simulator(*this);
         stokes_matrix_free->parse_parameters(prm);
         stokes_matrix_free->initialize();
       }
-
-    if (parameters.stokes_solver_type == Parameters<dim>::StokesSolverType::direct_solver)
+    else if (parameters.stokes_solver_type == Parameters<dim>::StokesSolverType::direct_solver)
       {
         stokes_direct = std::make_unique<StokesSolver::Direct<dim>>();
         stokes_direct->initialize_simulator(*this);
         stokes_direct->parse_parameters(prm);
         stokes_direct->initialize();
+      }
+    else
+      {
+        stokes_matrix_based = std::make_unique<StokesSolver::MatrixBased<dim>>(*this);
+        stokes_matrix_based->initialize_simulator(*this);
+        stokes_matrix_based->parse_parameters(prm);
+        stokes_matrix_based->initialize();
       }
 
     postprocess_manager.initialize_simulator (*this);
